% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorRamprgb}
\alias{colorRamprgb}
\title{Colour ramp vector}
\usage{
colorRamprgb(colour1, colour2, length)
}
\arguments{
\item{colour1}{valid argument for \code{col2rgb}}

\item{colour2}{valid argument for \code{col2rgb}}

\item{length}{integer determining legnth of returned vector}
}
\value{
A character vector of hex colours
}
\description{
Generate a colour ramp with hex vector output
}
\examples{
colorRamprgb(inky_colours("inky-red"), inky_colours("inky-blue"), 20)
}
