% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inky-gradients.R
\name{inky_2d_fig}
\alias{inky_2d_fig}
\title{Make a gradient figure from four colours}
\usage{
inky_2d_fig(colours, x = 20, y = 14)
}
\arguments{
\item{colours}{A vector of length 4, all as valid arguments to \code{col2rgb}}

\item{x}{The length of the x dimension}

\item{y}{The length of the y dimension}
}
\value{
A ggplot2 objects
}
\description{
Make a gradient figure from four colours
}
\examples{
inky_2d_fig(inky_colours()[3:6])
}
