% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inky-gradients.R
\name{inky_2d_space}
\alias{inky_2d_space}
\title{Make a 2D colour gradient}
\usage{
inky_2d_space(colours, x = 20, y = 14)
}
\arguments{
\item{colours}{A vector of length 4, all as valid arguments to \code{col2rgb}}

\item{x}{The length of the x dimension}

\item{y}{The length of the y dimension}
}
\value{
A data.frame
}
\description{
Make a 2D colour gradient
}
\examples{

grad_2d <- inky_2d_space(inky_colours()[3:6])
}
