% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{inky_colours}
\alias{inky_colours}
\title{Clour definitions}
\usage{
inky_colours(...)
}
\arguments{
\item{...}{Which colours to return}
}
\value{

}
\description{
Clour definitions
}
\examples{
inky_colours("inky-orange")
#inky-orange
   "#AC7B00"
}
