% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{inky_pal}
\alias{inky_pal}
\title{Interpolate an inky palette}
\usage{
inky_pal(palette = "inky_basic", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to colorRampPalette()}
}
\description{
Interpolate an inky palette
}
