% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{inky_palettes}
\alias{inky_palettes}
\title{Inky compatible colours}
\format{
An object of class \code{list} of length 5.
}
\usage{
inky_palettes
}
\description{
Inky compatible colours
}
\keyword{datasets}
