% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inky-ggsave.R
\name{inkysave}
\alias{inkysave}
\title{Use ggsave to export in a compatible way for inky impression}
\usage{
inkysave(plot, filename, ...)
}
\arguments{
\item{plot}{a ggplot object}

\item{filename}{path to save to}

\item{...}{other params passed to ggsave}
}
\description{
Use ggsave to export in a compatible way for inky impression
}
\examples{
df <- tidyr::crossing(x = 1:19, y = 1:16)
df$hex <- sample(factor(1:7), nrow(df), replace = TRUE)

ggplot(df, aes(x, y)) +
 geom_tile(aes(fill = hex)) +
 scale_fill_inky(guide = FALSE) +
 ggplot2::theme_void() +
 ggplot2::coord_cartesian(expand = FALSE) -> p
 inkysave(p, 'filename.jpg')
}
