% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_fill_inky}
\alias{scale_fill_inky}
\title{Construct a fill palette}
\usage{
scale_fill_inky(palette = "inky_basic", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Construct a fill palette
}
